Move Transaction classes to Phragile\Domain namespace
