+++ app/Phragile/ClosedTimeDispatcher.php
@@ -1,6 +1,8 @@
 <?php
 namespace Phragile;
 
+use Phragile\Domain\Transaction;
+
 interface ClosedTimeDispatcher {
 	public function isClosingTransaction(Transaction $transaction);
 }
