+++ app/Phragile/Domain/ColumnChangeTransaction.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phragile;
+namespace Phragile\Domain;
 
 class ColumnChangeTransaction extends Transaction {
