+++ app/Phragile/Domain/MergeAndCloseTransaction.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Phragile;
+namespace Phragile\Domain;
 
 // In Phabricator world when the task is merged into the other task (e.g. as it is considered a duplicate
 // of the other task), the merged task is considered closed (there is no status change transaction
 // changing status to "resolved").
+
 class MergeAndCloseTransaction extends Transaction {
 
 	const TYPE = 'mergeAndClose';
